Moves APIs and Calculator back out of views.py.
