+++ komparu/comparar/angelList.py
@@ -3,21 +3,18 @@ import urllib
 
 class AngelListAPI:
     def get_scores(self, a, b):
-        count_a = self.get_score(a)
-        count_b = self.get_score(b)
+        count_a = self.get_count(a)
+        count_b = self.get_count(b)
 
         total = count_a + count_b
 
-        if total == 0:
-            return {a: 0.0, b: 0.0}, "Angel List"
-
-        adjusted_ratio_a = count_a * 1.0 / total / 10
-        adjusted_ratio_b = count_b * 1.0 / total / 10
+        score_a = self.get_score(total, count_a)
+        score_b = self.get_score(total, count_b)
 
         # more startups means you're probably worse
-        return {a: 1.0 - adjusted_ratio_a, b: 1.0 - adjusted_ratio_b}, "Angel List"
+        return "Angel List", {a: score_a, b: score_b}, {a: count_a, b: count_b}
 
-    def get_score(self, item):
+    def get_count(self, item):
         response = unirest.get(
             "https://api.angel.co/1/search?query={0}&type=Startup".format(
                 urllib.quote_plus(item))
@@ -28,5 +25,11 @@ class AngelListAPI:
         if not results:
             return 0
         else:
+            # number of startups
             return len(results)
 
+    def get_score(self, total, count):
+        if total == 0:
+            return 1.0
+
+        return 1.0 - (count / total / 10.0)
